修复了大部分枚举常量缺失、属性类型错误等异常

+++ RhythmBase/RhythmDoctor/Events/PulseCamera.cs
@@ -12,7 +12,7 @@ public class PulseCamera : BaseEvent, IRoomEvent
 	/// <summary>
 	/// Gets or sets the strength of the pulse.
 	/// </summary>
-	public byte Strength { get; set; } = 1;
+	public int Strength { get; set; } = 1;
 	/// <summary>
 	/// Gets or sets the count of pulses.
 	/// </summary>
